Фикс проблемы с работой accessChecker. Запросы для работы с пинами доски перенесены в домен boards из домена pins.
