+++ internal/boards/delivery/http/delivery.go
@@ -18,10 +18,14 @@ func RegisterHandlers(mux *httprouter.Router, logger log.Logger, authorizer mw.A
 
 	mux.POST("/boards", mw.HandleLogger(mw.ErrorHandler(authorizer(mw.CorsChecker(del.create)), logger), logger))
 	mux.GET("/boards", mw.HandleLogger(mw.ErrorHandler(authorizer(mw.CorsChecker(del.list)), logger), logger))
-	mux.GET("/boards/:board_id", mw.HandleLogger(mw.ErrorHandler(mw.CorsChecker(authorizer(access.ReadChecker(del.get))), logger), logger))
-	mux.PUT("/boards/:board_id", mw.HandleLogger(mw.ErrorHandler(mw.CorsChecker(authorizer(access.WriteChecker(del.fullUpdate))), logger), logger))
-	mux.PATCH("/boards/:board_id", mw.HandleLogger(mw.ErrorHandler(mw.CorsChecker(authorizer(access.WriteChecker(del.partialUpdate))), logger), logger))
-	mux.DELETE("/boards/:board_id", mw.HandleLogger(mw.ErrorHandler(mw.CorsChecker(authorizer(access.WriteChecker(del.delete))), logger), logger))
+	mux.GET("/boards/:id", mw.HandleLogger(mw.ErrorHandler(mw.CorsChecker(authorizer(access.ReadChecker(del.get))), logger), logger))
+	mux.PUT("/boards/:id", mw.HandleLogger(mw.ErrorHandler(mw.CorsChecker(authorizer(access.WriteChecker(del.fullUpdate))), logger), logger))
+	mux.PATCH("/boards/:id", mw.HandleLogger(mw.ErrorHandler(mw.CorsChecker(authorizer(access.WriteChecker(del.partialUpdate))), logger), logger))
+	mux.DELETE("/boards/:id", mw.HandleLogger(mw.ErrorHandler(mw.CorsChecker(authorizer(access.WriteChecker(del.delete))), logger), logger))
+
+	mux.POST("/boards/:id/pins/:pin_id", mw.HandleLogger(mw.ErrorHandler(authorizer(mw.CorsChecker(access.WriteChecker(del.addPin))), logger), logger))
+	mux.GET("/boards/:id/pins", mw.HandleLogger(mw.ErrorHandler(authorizer(mw.CorsChecker(access.ReadChecker(del.pinsList))), logger), logger))
+	mux.DELETE("/boards/:id/pins/:pin_id", mw.HandleLogger(mw.ErrorHandler(authorizer(mw.CorsChecker(access.WriteChecker(del.removePin))), logger), logger))
 }
 
 type delivery struct {
@@ -96,7 +100,7 @@ func (del *delivery) list(w http.ResponseWriter, r *http.Request, p httprouter.P
 }
 
 func (del *delivery) get(w http.ResponseWriter, r *http.Request, p httprouter.Params) error {
-	strId := p.ByName("board_id")
+	strId := p.ByName("id")
 	id, err := strconv.Atoi(strId)
 	if err != nil {
 		return mw.ErrInvalidBoardIdParam
@@ -119,12 +123,13 @@ func (del *delivery) get(w http.ResponseWriter, r *http.Request, p httprouter.Pa
 		return mw.ErrCreateResponse
 	}
 
+	w.Header().Set("Content-Type", "application/json")
 	_, err = w.Write(data)
 	return err
 }
 
 func (del *delivery) fullUpdate(w http.ResponseWriter, r *http.Request, p httprouter.Params) error {
-	strId := p.ByName("board_id")
+	strId := p.ByName("id")
 	id, err := strconv.Atoi(strId)
 	if err != nil {
 		return mw.ErrInvalidBoardIdParam
@@ -174,7 +179,7 @@ func (del *delivery) fullUpdate(w http.ResponseWriter, r *http.Request, p httpro
 }
 
 func (del *delivery) partialUpdate(w http.ResponseWriter, r *http.Request, p httprouter.Params) error {
-	strId := p.ByName("board_id")
+	strId := p.ByName("id")
 	id, err := strconv.Atoi(strId)
 	if err != nil {
 		return mw.ErrInvalidBoardIdParam
@@ -231,7 +236,7 @@ func (del *delivery) partialUpdate(w http.ResponseWriter, r *http.Request, p htt
 }
 
 func (del *delivery) delete(w http.ResponseWriter, r *http.Request, p httprouter.Params) error {
-	strId := p.ByName("board_id")
+	strId := p.ByName("id")
 	id, err := strconv.Atoi(strId)
 	if err != nil {
 		return mw.ErrInvalidBoardIdParam
@@ -244,6 +249,92 @@ func (del *delivery) delete(w http.ResponseWriter, r *http.Request, p httprouter
 		} else {
 			return mw.ErrService
 		}
+	}
+	return nil
+}
+
+func (del *delivery) pinsList(w http.ResponseWriter, r *http.Request, p httprouter.Params) error {
+	strBoardId := p.ByName("id")
+	boardId, err := strconv.Atoi(strBoardId)
+	if err != nil {
+		return mw.ErrInvalidBoardIdParam
+	}
+
+	queryValues := r.URL.Query()
+	page := 1
+	strPage := queryValues.Get("page")
+	if strPage != "" {
+		page, err = strconv.Atoi(strPage)
+		if err != nil {
+			return mw.ErrInvalidPageParam
+		} else if page < 1 {
+			return mw.ErrInvalidPageParam
+		}
+	}
+
+	limit := 30
+	strLimit := queryValues.Get("limit")
+	if strLimit != "" {
+		limit, err = strconv.Atoi(strLimit)
+		if err != nil {
+			return mw.ErrInvalidLimitParam
+		} else if limit < 0 {
+			return mw.ErrInvalidLimitParam
+		}
+	}
+
+	pins, err := del.serv.PinsList(boardId, page, limit)
+	if err != nil {
+		return mw.ErrService
+	}
+
+	response := pinListResponse{Pins: pins}
+	data, err := json.Marshal(response)
+	if err != nil {
+		return mw.ErrCreateResponse
+	}
+
+	w.Header().Set("Content-Type", "application/json")
+	_, err = w.Write(data)
+	return err
+}
+
+func (del *delivery) addPin(w http.ResponseWriter, r *http.Request, p httprouter.Params) error {
+	strId := p.ByName("id")
+	boardId, err := strconv.Atoi(strId)
+	if err != nil {
+		return mw.ErrInvalidBoardIdParam
+	}
+
+	strId = p.ByName("pin_id")
+	pinId, err := strconv.Atoi(strId)
+	if err != nil {
+		return mw.ErrInvalidPinIdParam
+	}
+
+	err = del.serv.AddPin(boardId, pinId)
+	if err != nil {
+		return mw.ErrService
+	}
+	return nil
+}
+
+func (del *delivery) removePin(w http.ResponseWriter, r *http.Request, p httprouter.Params) error {
+	strId := p.ByName("id")
+	boardId, err := strconv.Atoi(strId)
+	if err != nil {
+		return mw.ErrInvalidBoardIdParam
+	}
+
+	strId = p.ByName("pin_id")
+	pinId, err := strconv.Atoi(strId)
+	if err != nil {
+		return mw.ErrInvalidPinIdParam
+	}
+
+	err = del.serv.RemovePin(boardId, pinId)
+	if err != nil {
+		return mw.ErrService
 	}
 	return nil
 }
