+++ internal/boards/delivery/http/delivery_test.go
@@ -1,6 +1,7 @@
 package http
 
 import (
+	"github.com/go-park-mail-ru/2023_1_PracticalDev/internal/pins"
 	"io"
 	"net/http"
 	"net/http/httptest"
@@ -197,13 +198,13 @@ func TestGet(t *testing.T) {
 					UserId:      1,
 				}, nil)
 			},
-			params:   []httprouter.Param{{Key: "board_id", Value: "3"}},
+			params:   []httprouter.Param{{Key: "id", Value: "3"}},
 			response: `{"id":3,"name":"n3","description":"d3","privacy":"secret","user_id":1}`,
 			err:      nil,
 		},
 		"invalid board id param": {
 			prepare:  func(f *fields) {},
-			params:   []httprouter.Param{{Key: "board_id", Value: "a"}},
+			params:   []httprouter.Param{{Key: "id", Value: "a"}},
 			response: ``,
 			err:      mw.ErrInvalidBoardIdParam,
 		},
@@ -217,7 +218,7 @@ func TestGet(t *testing.T) {
 			prepare: func(f *fields) {
 				f.serv.EXPECT().Get(3).Return(models.Board{}, _boards.ErrBoardNotFound)
 			},
-			params:   []httprouter.Param{{Key: "board_id", Value: "3"}},
+			params:   []httprouter.Param{{Key: "id", Value: "3"}},
 			response: ``,
 			err:      mw.ErrBoardNotFound,
 		},
@@ -284,14 +285,14 @@ func TestFullUpdate(t *testing.T) {
 					Privacy:     "secret",
 					UserId:      3}, nil)
 			},
-			params:   []httprouter.Param{{Key: "board_id", Value: "1"}},
+			params:   []httprouter.Param{{Key: "id", Value: "1"}},
 			request:  `{"name":"b1","description":"d1","privacy":"secret"}`,
 			response: `{"id":1,"name":"b1","description":"d1","privacy":"secret","user_id":3}`,
 			err:      nil,
 		},
 		"invalid board id param": {
 			prepare:  func(f *fields) {},
-			params:   []httprouter.Param{{Key: "board_id", Value: "a"}},
+			params:   []httprouter.Param{{Key: "id", Value: "a"}},
 			response: ``,
 			err:      mw.ErrInvalidBoardIdParam,
 		},
@@ -367,14 +368,14 @@ func TestPartialUpdate(t *testing.T) {
 					Privacy:     "secret",
 					UserId:      3}, nil)
 			},
-			params:   []httprouter.Param{{Key: "board_id", Value: "1"}},
+			params:   []httprouter.Param{{Key: "id", Value: "1"}},
 			request:  `{"name":"b1","description":"d1","privacy":"secret"}`,
 			response: `{"id":1,"name":"b1","description":"d1","privacy":"secret","user_id":3}`,
 			err:      nil,
 		},
 		"invalid board id param": {
 			prepare:  func(f *fields) {},
-			params:   []httprouter.Param{{Key: "board_id", Value: "a"}},
+			params:   []httprouter.Param{{Key: "id", Value: "a"}},
 			response: ``,
 			err:      mw.ErrInvalidBoardIdParam,
 		},
@@ -435,12 +436,12 @@ func TestDelete(t *testing.T) {
 			prepare: func(f *fields) {
 				f.serv.EXPECT().Delete(3).Return(nil)
 			},
-			params: []httprouter.Param{{Key: "board_id", Value: "3"}},
+			params: []httprouter.Param{{Key: "id", Value: "3"}},
 			err:    nil,
 		},
 		"invalid board id param": {
 			prepare: func(f *fields) {},
-			params:  []httprouter.Param{{Key: "board_id", Value: "a"}},
+			params:  []httprouter.Param{{Key: "id", Value: "a"}},
 			err:     mw.ErrInvalidBoardIdParam,
 		},
 		"missing board id param": {
@@ -452,7 +453,7 @@ func TestDelete(t *testing.T) {
 			prepare: func(f *fields) {
 				f.serv.EXPECT().Delete(3).Return(_boards.ErrBoardNotFound)
 			},
-			params: []httprouter.Param{{Key: "board_id", Value: "3"}},
+			params: []httprouter.Param{{Key: "id", Value: "3"}},
 			err:    mw.ErrBoardNotFound,
 		},
 	}
@@ -481,6 +482,162 @@ func TestDelete(t *testing.T) {
 			err := del.delete(rec, req, test.params)
 			if err != test.err {
 				t.Errorf("\nExpected: %s\nGot: %s", test.err, err)
+			}
+		})
+	}
+}
+
+func TestAddPin(t *testing.T) {
+	type fields struct {
+		serv *mocks.MockService
+	}
+
+	type testCase struct {
+		prepare func(f *fields)
+		params  httprouter.Params
+		request string
+		err     error
+	}
+
+	tests := map[string]testCase{
+		"usual": {
+			prepare: func(f *fields) {
+				f.serv.EXPECT().AddPin(3, 2).Return(nil)
+			},
+			params: []httprouter.Param{
+				{Key: "id", Value: "3"},
+				{Key: "pin_id", Value: "2"},
+			},
+			err: nil,
+		},
+		"invalid board id param": {
+			prepare: func(f *fields) {},
+			params: []httprouter.Param{
+				{Key: "id", Value: "a"},
+				{Key: "pin_id", Value: "3"},
+			},
+			err: mw.ErrInvalidBoardIdParam,
+		},
+		"invalid pin id": {
+			prepare: func(f *fields) {},
+			params: []httprouter.Param{
+				{Key: "id", Value: "3"},
+				{Key: "pin_id", Value: "a"},
+			},
+			err: mw.ErrInvalidPinIdParam,
+		},
+		"service error": {
+			prepare: func(f *fields) {
+				f.serv.EXPECT().AddPin(3, 2).Return(pins.ErrDb)
+			},
+			params: []httprouter.Param{
+				{Key: "id", Value: "3"},
+				{Key: "pin_id", Value: "2"},
+			},
+			err: mw.ErrService,
+		},
+	}
+
+	for name, test := range tests {
+		test := test
+		t.Run(name, func(t *testing.T) {
+			t.Parallel()
+
+			ctrl := gomock.NewController(t)
+			defer ctrl.Finish()
+
+			f := fields{serv: mocks.NewMockService(ctrl)}
+			if test.prepare != nil {
+				test.prepare(&f)
+			}
+
+			logger := log.New()
+			del := delivery{
+				serv: f.serv,
+				log:  logger,
+			}
+
+			req := httptest.NewRequest(http.MethodPost, "/boards/3/pins/2", strings.NewReader(test.request))
+			rec := httptest.NewRecorder()
+			err := del.addPin(rec, req, test.params)
+			if err != test.err {
+				t.Errorf("\nExpected: %s\nGot: %s", test.err, err)
+			}
+		})
+	}
+}
+
+func TestPinsList(t *testing.T) {
+	type fields struct {
+		serv *mocks.MockService
+	}
+
+	type testCase struct {
+		prepare  func(f *fields)
+		params   httprouter.Params
+		response string
+		err      error
+	}
+
+	tests := map[string]testCase{
+		"usual": {
+			prepare: func(f *fields) {
+				f.serv.EXPECT().PinsList(12, 1, 30).Return([]models.Pin{
+					{Id: 1, Title: "t1", MediaSource: "ms_url1", Description: "d1", Author: 12},
+					{Id: 2, Title: "t2", MediaSource: "ms_url2", Description: "d2", Author: 10},
+					{Id: 3, Title: "t3", MediaSource: "ms_url3", Description: "d3", Author: 3},
+				}, nil)
+			},
+			params: []httprouter.Param{
+				{Key: "page", Value: "1"},
+				{Key: "limit", Value: "30"},
+				{Key: "id", Value: "12"},
+			},
+			response: `{"pins":[{"id":1,"title":"t1","description":"d1","media_source":"ms_url1","author_id":12},{"id":2,"title":"t2","description":"d2","media_source":"ms_url2","author_id":10},{"id":3,"title":"t3","description":"d3","media_source":"ms_url3","author_id":3}]}`,
+			err:      nil,
+		},
+		"no pins": {
+			prepare: func(f *fields) {
+				f.serv.EXPECT().PinsList(12, 1, 30).Return([]models.Pin{}, nil)
+			},
+			params: []httprouter.Param{
+				{Key: "page", Value: "1"},
+				{Key: "limit", Value: "30"},
+				{Key: "id", Value: "12"},
+			},
+			response: `{"pins":[]}`,
+			err:      nil,
+		},
+	}
+
+	for name, test := range tests {
+		test := test
+		t.Run(name, func(t *testing.T) {
+			t.Parallel()
+
+			ctrl := gomock.NewController(t)
+			defer ctrl.Finish()
+
+			f := fields{serv: mocks.NewMockService(ctrl)}
+			if test.prepare != nil {
+				test.prepare(&f)
+			}
+
+			logger := log.New()
+			del := delivery{
+				serv: f.serv,
+				log:  logger,
+			}
+
+			req := httptest.NewRequest(http.MethodGet, "/boards/12/pins", nil)
+			rec := httptest.NewRecorder()
+			err := del.pinsList(rec, req, test.params)
+			if err != test.err {
+				t.Errorf("\nExpected: %s\nGot: %s", test.err, err)
+			}
+			body, _ := io.ReadAll(rec.Body)
+			if strings.Trim(string(body), "\n") != test.response {
+				t.Errorf("\nExpected: %s\nGot: %s", test.response, string(body))
 			}
 		})
 	}
