+++ internal/boards/repository.go
@@ -43,6 +43,10 @@ type Repository interface {
 	PartialUpdate(params *PartialUpdateParams) (models.Board, error)
 	Delete(id int) error
 
+	AddPin(boardId, pinId int) error
+	PinsList(boardId int, page, limit int) ([]models.Pin, error)
+	RemovePin(boardId, pinId int) error
+
 	CheckWriteAccess(userId, boardId string) (bool, error)
 	CheckReadAccess(userId, boardId string) (bool, error)
 }
