+++ internal/boards/service/service.go
@@ -5,12 +5,12 @@ import (
 	"github.com/go-park-mail-ru/2023_1_PracticalDev/internal/models"
 )
 
-type boardsService struct {
+type service struct {
 	repo boards.Repository
 }
 
 func NewBoardsService(repo boards.Repository) boards.Service {
-	return &boardsService{repo: repo}
+	return &service{repo: repo}
 }
 
 func validatePrivacy(privacy string) error {
@@ -20,37 +20,49 @@ func validatePrivacy(privacy string) error {
 	return nil
 }
 
-func (serv *boardsService) Create(params *boards.CreateParams) (models.Board, error) {
+func (serv *service) Create(params *boards.CreateParams) (models.Board, error) {
 	if err := validatePrivacy(params.Privacy); err != nil {
 		return models.Board{}, err
 	}
 	return serv.repo.Create(params)
 }
 
-func (serv *boardsService) List(userId int) ([]models.Board, error) {
+func (serv *service) List(userId int) ([]models.Board, error) {
 	return serv.repo.List(userId)
 }
 
-func (serv *boardsService) Get(id int) (models.Board, error) {
+func (serv *service) Get(id int) (models.Board, error) {
 	return serv.repo.Get(id)
 }
 
-func (serv *boardsService) FullUpdate(params *boards.FullUpdateParams) (models.Board, error) {
+func (serv *service) FullUpdate(params *boards.FullUpdateParams) (models.Board, error) {
 	return serv.repo.FullUpdate(params)
 }
 
-func (serv *boardsService) PartialUpdate(params *boards.PartialUpdateParams) (models.Board, error) {
+func (serv *service) PartialUpdate(params *boards.PartialUpdateParams) (models.Board, error) {
 	return serv.repo.PartialUpdate(params)
 }
 
-func (serv *boardsService) Delete(id int) error {
+func (serv *service) Delete(id int) error {
 	return serv.repo.Delete(id)
 }
 
-func (serv *boardsService) CheckWriteAccess(userId, boardId string) (bool, error) {
+func (serv *service) AddPin(boardId, pinId int) error {
+	return serv.repo.AddPin(boardId, pinId)
+}
+
+func (serv *service) PinsList(boardId int, page, limit int) ([]models.Pin, error) {
+	return serv.repo.PinsList(boardId, page, limit)
+}
+
+func (serv *service) RemovePin(boardId, pinId int) error {
+	return serv.repo.RemovePin(boardId, pinId)
+}
+
+func (serv *service) CheckWriteAccess(userId, boardId string) (bool, error) {
 	return serv.repo.CheckWriteAccess(userId, boardId)
 }
 
-func (serv *boardsService) CheckReadAccess(userId, boardId string) (bool, error) {
+func (serv *service) CheckReadAccess(userId, boardId string) (bool, error) {
 	return serv.repo.CheckReadAccess(userId, boardId)
 }
