+++ internal/pins/delivery/http/delivery_test.go
@@ -260,82 +260,6 @@ func TestListByUser(t *testing.T) {
 	}
 }
 
-func TestListByBoard(t *testing.T) {
-	type fields struct {
-		serv *mocks.MockService
-	}
-
-	type testCase struct {
-		prepare  func(f *fields)
-		params   httprouter.Params
-		response string
-		err      error
-	}
-
-	tests := map[string]testCase{
-		"usual": {
-			prepare: func(f *fields) {
-				f.serv.EXPECT().ListByBoard(12, 1, 30).Return([]models.Pin{
-					{Id: 1, Title: "t1", MediaSource: "ms_url1", Description: "d1", Author: 12},
-					{Id: 2, Title: "t2", MediaSource: "ms_url2", Description: "d2", Author: 10},
-					{Id: 3, Title: "t3", MediaSource: "ms_url3", Description: "d3", Author: 3},
-				}, nil)
-			},
-			params: []httprouter.Param{
-				{Key: "page", Value: "1"},
-				{Key: "limit", Value: "30"},
-				{Key: "board_id", Value: "12"},
-			},
-			response: `{"pins":[{"id":1,"title":"t1","description":"d1","media_source":"ms_url1","author_id":12},{"id":2,"title":"t2","description":"d2","media_source":"ms_url2","author_id":10},{"id":3,"title":"t3","description":"d3","media_source":"ms_url3","author_id":3}]}`,
-			err:      nil,
-		},
-		"no pins": {
-			prepare: func(f *fields) {
-				f.serv.EXPECT().ListByBoard(12, 1, 30).Return([]models.Pin{}, nil)
-			},
-			params: []httprouter.Param{
-				{Key: "page", Value: "1"},
-				{Key: "limit", Value: "30"},
-				{Key: "board_id", Value: "12"},
-			},
-			response: `{"pins":[]}`,
-			err:      nil,
-		},
-	}
-
-	for name, test := range tests {
-		test := test
-		t.Run(name, func(t *testing.T) {
-			t.Parallel()
-
-			ctrl := gomock.NewController(t)
-			defer ctrl.Finish()
-
-			f := fields{serv: mocks.NewMockService(ctrl)}
-			if test.prepare != nil {
-				test.prepare(&f)
-			}
-
-			logger := log.New()
-			del := delivery{
-				serv: f.serv,
-				log:  logger,
-			}
-
-			req := httptest.NewRequest(http.MethodGet, "/boards/12/pins", nil)
-			rec := httptest.NewRecorder()
-			err := del.listByBoard(rec, req, test.params)
-			if err != test.err {
-				t.Errorf("\nExpected: %s\nGot: %s", test.err, err)
-			}
-			body, _ := io.ReadAll(rec.Body)
-			if strings.Trim(string(body), "\n") != test.response {
-				t.Errorf("\nExpected: %s\nGot: %s", test.response, string(body))
-			}
-		})
-	}
-}
-
 func TestGet(t *testing.T) {
 	type fields struct {
 		serv *mocks.MockService
@@ -593,86 +517,6 @@ func TestDelete(t *testing.T) {
 			req := httptest.NewRequest(http.MethodDelete, "/pins/3", nil)
 			rec := httptest.NewRecorder()
 			err := del.delete(rec, req, test.params)
-			if err != test.err {
-				t.Errorf("\nExpected: %s\nGot: %s", test.err, err)
-			}
-		})
-	}
-}
-
-func TestAddToBoard(t *testing.T) {
-	type fields struct {
-		serv *mocks.MockService
-	}
-
-	type testCase struct {
-		prepare func(f *fields)
-		params  httprouter.Params
-		request string
-		err     error
-	}
-
-	tests := map[string]testCase{
-		"usual": {
-			prepare: func(f *fields) {
-				f.serv.EXPECT().AddToBoard(3, 2).Return(nil)
-			},
-			params: []httprouter.Param{
-				{Key: "board_id", Value: "3"},
-				{Key: "id", Value: "2"},
-			},
-			err: nil,
-		},
-		"invalid board id param": {
-			prepare: func(f *fields) {},
-			params: []httprouter.Param{
-				{Key: "board_id", Value: "a"},
-				{Key: "id", Value: "3"},
-			},
-			err: mw.ErrInvalidBoardIdParam,
-		},
-		"invalid pin id": {
-			prepare: func(f *fields) {},
-			params: []httprouter.Param{
-				{Key: "board_id", Value: "3"},
-				{Key: "id", Value: "a"},
-			},
-			err: mw.ErrInvalidPinIdParam,
-		},
-		"service error": {
-			prepare: func(f *fields) {
-				f.serv.EXPECT().AddToBoard(3, 2).Return(pins.ErrDb)
-			},
-			params: []httprouter.Param{
-				{Key: "board_id", Value: "3"},
-				{Key: "id", Value: "2"},
-			},
-			err: mw.ErrService,
-		},
-	}
-
-	for name, test := range tests {
-		test := test
-		t.Run(name, func(t *testing.T) {
-			t.Parallel()
-
-			ctrl := gomock.NewController(t)
-			defer ctrl.Finish()
-
-			f := fields{serv: mocks.NewMockService(ctrl)}
-			if test.prepare != nil {
-				test.prepare(&f)
-			}
-
-			logger := log.New()
-			del := delivery{
-				serv: f.serv,
-				log:  logger,
-			}
-
-			req := httptest.NewRequest(http.MethodPost, "/boards/3/pins/2", strings.NewReader(test.request))
-			rec := httptest.NewRecorder()
-			err := del.addToBoard(rec, req, test.params)
 			if err != test.err {
 				t.Errorf("\nExpected: %s\nGot: %s", test.err, err)
 			}
