+++ internal/pins/mocks/service.go
@@ -35,20 +35,6 @@ func (m *MockService) EXPECT() *MockServiceMockRecorder {
 	return m.recorder
 }
 
-// AddToBoard mocks base method.
-func (m *MockService) AddToBoard(boardId, pinId int) error {
-	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "AddToBoard", boardId, pinId)
-	ret0, _ := ret[0].(error)
-	return ret0
-}
-
-// AddToBoard indicates an expected call of AddToBoard.
-func (mr *MockServiceMockRecorder) AddToBoard(boardId, pinId interface{}) *gomock.Call {
-	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "AddToBoard", reflect.TypeOf((*MockService)(nil).AddToBoard), boardId, pinId)
-}
-
 // CheckReadAccess mocks base method.
 func (m *MockService) CheckReadAccess(userId, pinId string) (bool, error) {
 	m.ctrl.T.Helper()
@@ -153,21 +139,6 @@ func (mr *MockServiceMockRecorder) List(page, limit interface{}) *gomock.Call {
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "List", reflect.TypeOf((*MockService)(nil).List), page, limit)
 }
 
-// ListByBoard mocks base method.
-func (m *MockService) ListByBoard(boardId, page, limit int) ([]models.Pin, error) {
-	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "ListByBoard", boardId, page, limit)
-	ret0, _ := ret[0].([]models.Pin)
-	ret1, _ := ret[1].(error)
-	return ret0, ret1
-}
-
-// ListByBoard indicates an expected call of ListByBoard.
-func (mr *MockServiceMockRecorder) ListByBoard(boardId, page, limit interface{}) *gomock.Call {
-	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "ListByBoard", reflect.TypeOf((*MockService)(nil).ListByBoard), boardId, page, limit)
-}
-
 // ListByUser mocks base method.
 func (m *MockService) ListByUser(userId, page, limit int) ([]models.Pin, error) {
 	m.ctrl.T.Helper()
@@ -181,18 +152,4 @@ func (m *MockService) ListByUser(userId, page, limit int) ([]models.Pin, error) 
 func (mr *MockServiceMockRecorder) ListByUser(userId, page, limit interface{}) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "ListByUser", reflect.TypeOf((*MockService)(nil).ListByUser), userId, page, limit)
-}
-
-// RemoveFromBoard mocks base method.
-func (m *MockService) RemoveFromBoard(boardId, pinId int) error {
-	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "RemoveFromBoard", boardId, pinId)
-	ret0, _ := ret[0].(error)
-	return ret0
-}
-
-// RemoveFromBoard indicates an expected call of RemoveFromBoard.
-func (mr *MockServiceMockRecorder) RemoveFromBoard(boardId, pinId interface{}) *gomock.Call {
-	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "RemoveFromBoard", reflect.TypeOf((*MockService)(nil).RemoveFromBoard), boardId, pinId)
 }
